move client utility functions
